refactor: split up methods into separate files
